+++ ui/app/components/send_/send-content/send-gas-row/send-gas-row.selectors.js
@@ -1,0 +1,9 @@
+const selectors = {
+  sendGasIsInError,
+}
+
+module.exports = selectors
+
+function sendGasIsInError (state) {
+  return state.send.errors.gasLoading
+}